smart-data-explorer/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
openai==1.3.0
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
plotly==5.18.0
scikit-learn==1.3.2
statsmodels==0.14.0
joblib==1.3.2